+++ src/entities/user/ui/index.ts
@@ -1,1 +1,2 @@
 export { UserDialog } from "./UserDialog.tsx"
+export { OpenUserDialogButton } from "./OpenUserDialogButton.tsx"
