+++ src/features/comments/ui/CommentsList.tsx
@@ -8,7 +8,7 @@ import { newCommentAtom } from "../../addComment/model"
 import { updateCommentAtom } from "../../updateComment/model"
 import { useDeleteCommentMutation } from "../../deleteComment/api"
 import { useLikeCommentMutation } from "../../likeComment/api"
-import { searchQueryAtom } from "../../serchPost/model"
+import { searchQueryAtom } from "../../searchPost/model"
 import { selectedPostAtom } from "../../postDetail/model"
 import { highlightText } from "../../../shared/lib"
 import { dialogAtomFamily } from "../../../shared/model"
