+++ src/features/deletePost/ui/index.ts
@@ -1,0 +1,1 @@
+export { DeletePostButton } from "./DeletePostButton.tsx"
