+++ src/features/pagination/ui/Pagination.tsx
@@ -1,6 +1,6 @@
 import { useAtom, useAtomValue } from "jotai"
 
-import { limitAtom, skipAtom } from "../../serchPost/model"
+import { limitAtom, skipAtom } from "../../searchPost/model"
 import { postsTotalAtom } from "../../postsWithUsers/model"
 import { Button } from "../../../shared/ui/common"
 import { SelectContainer, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../../shared/ui/select"
