+++ src/features/postDetail/ui/OpenPostDetailButton.tsx
@@ -1,0 +1,28 @@
+import { useSetAtom } from "jotai"
+import { Button } from "../../../shared/ui/common"
+import { dialogAtomFamily } from "../../../shared/model"
+import { selectedPostAtom } from "../model"
+
+import type { ReactNode } from "react"
+import type { PostType } from "../../../entities/post/model"
+
+interface Props {
+  children: ReactNode
+  post: PostType
+}
+
+export const OpenPostDetailButton = ({ children, post }: Props) => {
+  const setShowDetailDialog = useSetAtom(dialogAtomFamily("post-detail"))
+  const setSelectedPost = useSetAtom(selectedPostAtom)
+
+  const openPostDetail = (post: PostType) => {
+    setSelectedPost(post)
+    setShowDetailDialog(true)
+  }
+
+  return (
+    <Button variant="ghost" size="sm" onClick={() => openPostDetail(post)}>
+      {children}
+    </Button>
+  )
+}
