+++ src/features/postDetail/ui/index.ts
@@ -1,1 +1,2 @@
 export { PostDetailDialog } from "./PostDetailDialog.tsx"
+export { OpenPostDetailButton } from "./OpenPostDetailButton.tsx"
