+++ src/features/searchPost/lib/index.ts
