+++ src/features/searchPost/model/index.ts
