+++ src/features/searchPost/model/searchParams.type.ts
