+++ src/features/searchPost/ui/SearchInput.tsx
