+++ src/features/searchPost/ui/SearchPostBar.tsx
