+++ src/features/searchPost/ui/index.ts
