+++ src/features/updatePost/ui/index.ts
@@ -1,1 +1,2 @@
 export { UpdatePostDialog } from "./UpdatePostDialog.tsx"
+export { OpenUpdatePostDialogButton } from "./OpenUpdatePostDialogButton.tsx"
